tigerstyle: fix nostr-relay clock, collection, and result-discard warnings
- Extract monotonic_now() clock boundary helper in rate_limit.rs
- Add Vec::with_capacity() for 6 collection growth sites in storage.rs
- Replace let _ = with drop() for 6 result discards in relay.rs
- Collapse nested if-let chains in kv_store.rs forwarding logic
- Use match guards in upstream_cache.rs for collapsible matches
- Promote debug_assert on constants to const assert across handler crates
- Fix usize->u32 in testing tester disconnect/connect wrappers
- Fix clippy: unnecessary cast, into_iter, absurd comparison, result_large_err

+++ crates/aspen-raft/src/log_subscriber.rs
@@ -630,7 +630,7 @@ async fn handle_log_subscriber_replay_batch(input: ReplayBatchInput<'_>) -> anyh
         send,
     } = input;
     debug_assert!(subscriber_id > 0, "subscriber_id must be positive during replay batch");
-    debug_assert!(entries.len() <= MAX_HISTORICAL_BATCH_SIZE as usize, "historical replay batch must stay bounded");
+    debug_assert!(entries.len() <= MAX_HISTORICAL_BATCH_SIZE, "historical replay batch must stay bounded");
 
     for entry in entries {
         *current_start = entry.index.saturating_add(1);
